Code to make the unit tests pass
This reads in the content of the CSV and stores each row as a hashref where the
identifier is the `track_upc`. Potentially the id makes more sense, but the
track_upc is also unique and allows for selecting by that rather than iterating
through all rows each time.

The unit tests all pass, though I do not handle the empty string properly so
there is a warning.